requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
google-genai